Don't re-write root logger log level
Currently the way that constants.py installs coloredlogs causes it to override the root logger to DEBUG, as well as enabling coloredlogs on the root logger, which can be undesirable. To fix that, instead we pass the (already created) logger instance into the install command, so that only log levels and coloring for that logger are modified.